% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeDegTable.R
\name{completeDegTtable}
\alias{completeDegTtable}
\title{Analyse a table of differentially expressed genes.}
\usage{
completeDegTtable(deg.table, table.name, sort.column = "none",
  thresholds = thresholds, round.digits = 3, dir.figures = NULL)
}
\arguments{
\item{deg.table}{Table with the results of differential expression analysis
obtained from RNA-seq data.
Such tables can be produced by a variety of packages, e.g. DESeq2, edgeR, etc.
However, the column names should be redefined since each of this packages uses
different names for similar metrics.
The required columns are c("gene.id", mean", "log2FC", "pvalue", "padj").
Additional columns can be provided but will be ignored for the analysis.}

\item{table.name}{name for the data table, which will be displayed in the plots.}

\item{sort.column="none"}{Column to sort the result.
Supported: c("none", "mean", "log2FC", "pvalue", "padj").}

\item{thresholds=c("pvalue"}{= 0.05, "padj"=0.05, "evalue"=1, "FC"=1.5)
Thresholds on some specific scores, used for display purpose, and to
add some columns to the result table, indicating if the gene passes the
threshold or not.}

\item{round.digits=3}{Significant digits to round the values of the
output table. Set to NA to avoid rounding.}

\item{dir.figures=NULL}{if not NULL, figures will be saved in the specified directory.}
}
\description{
Analyse a table of differentially expressed genes to
produce plots and additional statistics.
}
\details{
First version: 2015-08
Last modification: 2015-08.
}
\examples{

 deseq2.result.table <- data.frame(
    "gene.id" = row.names(deseq2.res),
    "mean" = deseq2.res$baseMean,
    "log2FC" = deseq2.res$log2FoldChange,
    "pvalue" = deseq2.res$pvalue,
    "padj" = deseq2.res$padj)
deseq2.result.table <- complete.deg.table(deseq2.result.table,
    table.name="DESeq2", sort.column="padj")

}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

