% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeEnrichTable.R
\name{completeEnrichTable}
\alias{completeEnrichTable}
\title{Add some columns to a functional enrichment table in order to ensure that we always have the same statistics.}
\usage{
completeEnrichTable(enrich.table, pvalue.column = "Pvalue",
  thresholds = c(evalue = 1, qvalue = 0.05), nb.tests = nrow(enrich.table),
  select.positives = FALSE, plot.adjust = TRUE)
}
\arguments{
\item{enrich.table}{Enrichment table from GOstats, clusterProfiler or similar programs.}

\item{pvalue.column}{Name or number of the column containing the p-values (this depends on the program).}
}
\description{
Add some columns to a functional enrichment table in order to ensure that we always have the same statistics.
}
\details{
First version: 2015-08
Last modification: 2015-08.
}
\examples{

}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

