% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalEnrichment.R
\name{functionalEnrichment}
\alias{functionalEnrichment}
\title{Run functional enrichment for a given gene set}
\usage{
functionalEnrichment(geneset, allgenes, db, organism.names, ontology = "BP",
  thresholds = thresholds, select.positives = TRUE, run.GOstats = TRUE,
  run.clusterProfiler = FALSE, plot.adjust = TRUE)
}
\arguments{
\item{geneset}{Entrez IDs of the gene set of interest.}

\item{allgenes}{EntrezIDs for all genes of the analyzed genome (the "universe").}

\item{db}{an annotation database}

\item{thresholds["evalue"]=0.1}{Upper threshold on the e-value for GOstats.
Note that we perform the selection with a p-value threshold of 1, in order to obtain the
stats for all the classes. The e-value is calculated a posteriori and indicated in the table.}

\item{select.positives=TRUE}{If TRUE, the result table only contains the positive tests,
i.e. those passing all the thresholds. Otherwise, all rows are returned, with a column "positive"
indicating whether or not they did or not pass all thresholds. Additional columns indicate the
result of the individual thresholds.}

\item{organism}{organism name in various tools. E.g:}
}
\description{
Run functional enrichment for a given gene set.
This function is a compilation of pieces of code from the vignettes
of different functional enrichment analysis package.
}
\details{
First version: 2015-08
Last modification: 2015-08.
}
\examples{

}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

