% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose.R
\name{verbose}
\alias{verbose}
\title{Display messages at a given verbosity level}
\usage{
verbose(message.content, level = 1, print.date = TRUE)
}
\arguments{
\item{verbosity}{Level of verbosity above which the message should be printed.}

\item{print.date=TRUE}{Print date and time}
}
\description{
Display messages depending on user-defined verbosity level
}
\details{
First version: 2015-03.
Last modification: 2015-03.
}
\examples{

verbosity <- 1 ## Define level of verbosity

## This message will be printed because the level is <= verbosity
verbose("This is printed", 1)

## This message will not be printed because the verbosity is inferior to the specified level
verbose("This is not printed", 2)

}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

